Add RestaurantDetail test
Add RestaurantDetail test.
