+++ fixtures/reviews.js
@@ -1,0 +1,16 @@
+const reviews = [
+  {
+    id: 1,
+    name: '테스터',
+    score: 3,
+    description: '맛있어요',
+  },
+  {
+    id: 2,
+    name: 'user1',
+    score: 4,
+    description: '짱맛있어요',
+  },
+];
+
+export default reviews;
