+++ src/ReviewList.test.jsx
@@ -2,22 +2,9 @@ import { render } from '@testing-library/react';
 
 import ReviewList from './ReviewList';
 
-describe('ReviewList', () => {
-  const reviews = [
-    {
-      id: 1,
-      name: '테스터',
-      score: 3,
-      description: '맛있어요',
-    },
-    {
-      id: 2,
-      name: 'user1',
-      score: 4,
-      description: '짱맛있어요',
-    },
-  ];
+import reviews from '../fixtures/reviews';
 
+describe('ReviewList', () => {
   it('renders reviewer name, score and description', () => {
     const { container } = render((
       <ReviewList reviews={reviews} />
